* 'develop' of https://gitlab.com/jusfoun-bigdata/hookah:
  客户同步信息
  日终文件上传通知
  日终文件上传通知
  日终文件上传通知
  修改xxx
  修改xxx
  补充提交...
  PaySign mongo
  fix bug
  发布商品BUG
  属性  类型更改
  fix bug
  字段调整
  出入金

+++ hookah-core/src/main/java/com/jusfoun/hookah/core/domain/bo/MoneyInOutBo.java
@@ -6,9 +6,9 @@ public class MoneyInOutBo {
 
     public Long payAccountID;
 
-    public Long Money;
+    public Long Money;          //  默认已经乘以100了
 
-    public Integer operatorType;
+    public byte operatorType;
 
     public String getUserId() {
         return userId;
@@ -34,11 +34,11 @@ public class MoneyInOutBo {
         Money = money;
     }
 
-    public Integer getOperatorType() {
+    public byte getOperatorType() {
         return operatorType;
     }
 
-    public void setOperatorType(Integer operatorType) {
+    public void setOperatorType(byte operatorType) {
         this.operatorType = operatorType;
     }
 }
